+++ outsystems/pipeline/deploy_package_to_target_env.py
@@ -22,7 +22,7 @@ from outsystems.vars.pipeline_vars import QUEUE_TIMEOUT_IN_SECS, SLEEP_PERIOD_IN
 # Functions
 from outsystems.lifetime.lifetime_environments import get_environment_key
 from outsystems.lifetime.lifetime_deployments import get_deployment_status, get_deployment_info, \
-    send_binary_deployment, delete_deployment, start_deployment, continue_deployment, get_running_deployment, \
+    send_binary_deployment, delete_deployment, start_deployment, continue_deployment, get_running_deployments, \
     check_deployment_two_step_deploy_status
 from outsystems.file_helpers.file import store_data, is_valid_os_package
 from outsystems.lifetime.lifetime_base import build_lt_endpoint
@@ -42,7 +42,7 @@ def main(artifact_dir: str, lt_http_proto: str, lt_url: str, lt_api_endpoint: st
 
     if not allow_parallel_deployments:
         wait_counter = 0
-        deployments = get_running_deployment(artifact_dir, lt_endpoint, lt_token, dest_env_key)
+        deployments = get_running_deployments(artifact_dir, lt_endpoint, lt_token, dest_env_key)
         while len(deployments) > 0:
             if wait_counter >= get_configuration_value("QUEUE_TIMEOUT_IN_SECS", QUEUE_TIMEOUT_IN_SECS):
                 print("Timeout occurred while waiting for LifeTime to be free, to create the new deployment plan.", flush=True)
@@ -51,7 +51,7 @@ def main(artifact_dir: str, lt_http_proto: str, lt_url: str, lt_api_endpoint: st
             sleep(sleep_value)
             wait_counter += sleep_value
             print("Waiting for LifeTime to be free. Elapsed time: {} seconds...".format(wait_counter), flush=True)
-            deployments = get_running_deployment(artifact_dir, lt_endpoint, lt_token, dest_env_key)
+            deployments = get_running_deployments(artifact_dir, lt_endpoint, lt_token, dest_env_key)
 
     # LT is free to deploy
     # Validate if file has OutSystems package extension
@@ -61,6 +61,9 @@ def main(artifact_dir: str, lt_http_proto: str, lt_url: str, lt_api_endpoint: st
     # Send the deployment plan and grab the key
     dep_plan_key = send_binary_deployment(artifact_dir, lt_endpoint, lt_token, lt_api_version, dest_env_key, package_path)
     print("Deployment plan {} created successfully.".format(dep_plan_key), flush=True)
+
+    # Write dep_plan_key to Azure DevOps variable
+    print(f"##vso[task.setvariable variable=DEPLOYMENT_PLAN_KEY;isOutput=true]{dep_plan_key}", flush=True)
 
     # Check if created deployment plan has conflicts
     dep_details = get_deployment_info(artifact_dir, lt_endpoint, lt_token, dep_plan_key)